+++ backend/main.py
@@ -1,7 +1,7 @@
-from Room import Room
-from Item import Item
+from room import Room
+from item import Item
 from character import Character
-from Messages import messages
+from messages import messages
 
 
 def handle_user_input(command):
