+++ server.py
@@ -2,7 +2,7 @@ from flask import Flask
 from flask import request
 from flask_cors import CORS
 import json
-from main import handle_user_input
+import backend
 
 app = Flask(__name__)
 CORS(app)
@@ -11,7 +11,7 @@ CORS(app)
 @app.route('/', methods=["GET", "POST"])
 def handle_request():
     command = str(request.args.get('command'))
-    output = handle_user_input(command)
+    output = backend.main.handle_user_input(command)
     data_set = {'output': output}
     json_dump = json.dumps(data_set)
     return json_dump
